Replace superfluous void Init() functions (MSW).
First step; replace the contents of the void Init() functions in places where they are used. Init conveys a meaning of "on startup" that wasn't always reflected in its usage and often we could just initialize in the constructor. The second part will move this initialization to the constructor declarations.