Replace wxArrayString uses with std::vector<wxString>. Remove overloads and narrow things down to accepting std::vector<wxString>. Minor formatting.
In the future, a better option for the interface here would be to use a span-like construct so any number of things can be passed in with ease and without worrying about making an excessive amount of overloads. While it is possible to use std/gsl::span right now, in most of the cases where wxArrayString has been replaced, we're taking ownership of the data.