*GetTextExtent functions now return a wxSize instead of having one pass ref / ptr parameters. Functions dealing that use double for extents return a pair of double.
Still isn't optimal, as the *GetTextExtent functions allow for retrieving a lot of aspects of text via ref / ptr params. A lot of instances have it where dummy parameters are passed in just to compile. In the future, this function will be broken apart so that it's explicit in getting what text properties you want.