Mark default constructors noexcept.
Some of the constructors are default and marked noexcept, but still may throw. They have defaulted parameters and should be broken up into a default constructor with no params and a constructor that takes params (but isn't noexcept). Splitting the two up also makes sense from the standpoint of if a constructor with params fails, we can use the default constructor instead.